requirements.txt
gspread
google-auth-oauthlib
google-auth
pandas
numpy
tqdm
requests
fastapi
uvicorn
openai
google-api-python-client